Uppdaterad

+++ src/main/java/org/example/HourlyPrice.java
@@ -1,15 +1,11 @@
 package org.example;
 
-import java.math.BigDecimal;
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-import java.util.Locale;
-
 public class HourlyPrice {
     private int startHour;
-    private BigDecimal price;
+    private int price;
 
-    public HourlyPrice(int startHour, BigDecimal price) {
+
+    public HourlyPrice(int startHour, int price) {
         this.startHour = startHour;
         this.price = price;
     }
@@ -22,7 +18,7 @@ public class HourlyPrice {
         return (startHour + 1) % 24;  // Calculate the end hour based on the start hour
     }
 
-    public BigDecimal getPrice() {
+    public int getPrice() {
         return price;
     }
 }
